+++ software/python-package/tests_manual/testbench_gen_fake.py
@@ -13,7 +13,7 @@ store_path = tmp_path / "harvest_example.h5"
 
 def random_data(length: int) -> np.ndarray:
     rng = np.random.default_rng()
-    return rng.integers(low=0, high=2**18, size=length, dtype="u4")
+    return rng.integers(low=0, high=2**18, size=length, dtype=np.uint32)
 
 
 with Writer(store_path, cal_data=CalibrationHarvester()) as store:
